call works with World league Method

+++ app/src/main/java/com/alphaomardiallo/duwagol/common/domain/model/prayerTimesCalendar/Params.kt
@@ -4,6 +4,6 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class Params(
-    val Fajr: Int,
-    val Isha: Int,
+    val Fajr: Double,
+    val Isha: Double,
 )
